requirements.txt
discord.py>=2.3.0
requests>=2.31.0
Pillow>=10.0.0
pyautogui>=0.9.54
pyperclip>=1.8.2
